requirements.txt
Flask==2.0.1
Flask-WTF==0.15.1
mysql-connector-python==8.0.26
requests==2.25.1